Moving BadRequestException from controllers.admin to controllers.
